requirements.txt
python-telegram-bot==13.15
redis==3.2.1
environs==14.3.0
requests==2.32.4